+++ sponsoredTransactions/backend/src/types.rs
@@ -34,21 +34,23 @@ pub enum LogError {
     AdditionalDataError,
     #[error("Node access error: {0}")]
     NodeAccess(#[from] QueryError),
-    #[error("Simulation of transaction rejected with reject reason: {0:?}")]
-    TransactionSimulationError(RejectReason),
+    /// The transaction simulation rejected where the exact
+    /// reject reason is unknown.
+    #[error("Simulation of transaction rejected. The exact reject reason is unknown: {0}")]
+    RejectedTransactionSimulation(#[from] UnknownDataError),
+    #[error("Simulation of transaction rejected with reject reason: {reason:?}")]
+    RejectedTransactionSimulationWithReason { reason: RejectReason },
     #[error(
         "Simulation of transaction rejected in smart contract with decoded reject reason: `{0}` \
          derived from: {1:?}."
     )]
-    TransactionSimulationDecodedError(DecodedReason, RejectReason),
+    RejectedTransactionSimulationWithDecodedReason(DecodedReason, RejectReason),
     #[error("Failed to create contract client: {0:?}")]
     FailedToCreateContractClient(QueryError),
     #[error("Invalid receive name: {0}")]
     ReceiveNameError(#[from] NewReceiveNameError),
     #[error("The input parameter exceeds the length limit: {0}")]
     ParameterSizeError(#[from] ExceedsParameterSize),
-    #[error("Unknown data error occured: {0}")]
-    UnknownDataError(#[from] UnknownDataError),
 }
 
 impl warp::reject::Reject for LogError {}