+++ trackAndTrace/indexer/src/bin/indexer.rs
@@ -113,9 +113,23 @@ impl indexer::ProcessEvent for StoreEvents {
             .context("Failed to execute latest_processed_block_height transaction")?;
 
         for single_contract_update_info in contract_update_info {
-            let contract_info = &single_contract_update_info.as_ref().known_or_err()?.0;
+            let contract_update_info = match single_contract_update_info.as_ref().known_or_err() {
+                Ok(details) => &details.0,
+                Err(e) => {
+                    // To reduce maintenance we ignore unknown contract update infos.
+                    tracing::warn!("Skipping indexing unknown contract update info: {e}");
+                    continue;
+                }
+            };
 
-            let tree = contract_info.execution_tree.as_ref().known_or_err()?;
+            let tree = match contract_update_info.execution_tree.as_ref().known_or_err() {
+                Ok(tree) => tree,
+                Err(e) => {
+                    // To reduce maintenance we ignore unknown smart contract execution trees.
+                    tracing::warn!("Skipping indexing unknown execution tree: {e}");
+                    continue;
+                }
+            };
 
             let smart_contract_events: Vec<(
                 ContractAddress,
@@ -123,8 +137,15 @@ impl indexer::ProcessEvent for StoreEvents {
                 &[ContractEvent],
             )> = tree
                 .events()
-                .map(|x| x.known_or_err())
-                .collect::<Result<_, _>>()?;
+                .filter_map(|event| match event.known_or_err() {
+                    Ok(event) => Some(event),
+                    Err(e) => {
+                        // To reduce maintenance we ignore unknown smart contract events.
+                        tracing::warn!("Skipping indexing unknown smart contract event: {e}");
+                        None
+                    }
+                })
+                .collect();
 
             for (_contract_invoked, _entry_point_name, events) in smart_contract_events {
                 for (event_index, event) in events.iter().enumerate() {
@@ -136,7 +157,7 @@ impl indexer::ProcessEvent for StoreEvents {
                     {
                         let params: [&(dyn ToSql + Sync); 6] = [
                             &(block_info.block_slot_time),
-                            &contract_info.transaction_hash.as_ref(),
+                            &contract_update_info.transaction_hash.as_ref(),
                             &(event_index as i64),
                             &(item_status_change_event.item_id.0 as i64),
                             &Json(&item_status_change_event.new_status),
@@ -162,7 +183,7 @@ impl indexer::ProcessEvent for StoreEvents {
                             "Preparing item_status_change_event from block {}, transaction hash \
                              {}, and event index {}.",
                             block_info.block_height,
-                            contract_info.transaction_hash,
+                            contract_update_info.transaction_hash,
                             event_index
                         );
                     } else if let contract::Event::<AdditionalData>::ItemCreated(
@@ -171,7 +192,7 @@ impl indexer::ProcessEvent for StoreEvents {
                     {
                         let params: [&(dyn ToSql + Sync); 6] = [
                             &(block_info.block_slot_time),
-                            &contract_info.transaction_hash.as_ref(),
+                            &contract_update_info.transaction_hash.as_ref(),
                             &(event_index as i64),
                             &(item_created_event.item_id.0 as i64),
                             &to_bytes(&item_created_event.metadata_url),
@@ -197,7 +218,7 @@ impl indexer::ProcessEvent for StoreEvents {
                             "Preparing event from block {}, transaction hash {}, and event index \
                              {}.",
                             block_info.block_height,
-                            contract_info.transaction_hash,
+                            contract_update_info.transaction_hash,
                             event_index
                         );
                     }