+++ sponsoredTransactionsAuction/backend/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased changes
 
-- Update `concordium-rust-sdk` dependency and adjust project to be forward-compatible.
-
+- Update `concordium-rust-sdk` dependency and adjust project to be forward-compatible. Unknown reject reasons when a sponsored transaction is simulated will produce an internal error.
+            
 ## 1.0.0
 
 - Initial auction sponsored transaction back end.