+++ trackAndTrace/indexer/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased changes
 
-- Update `rust-sdk` dependency so that project is forward compatible.
+- Update `rust-sdk` dependency so that project is forward compatible. Unknown contract update infos, 
+unkown smart contract events, and unkown smart contract execution trees are skipped when indexing.
 
 ## 0.2.1